contracts: Distinguish strictly between "script hash" and "code hash"
We want to use a script hash, whenever we want to compare scripts, as the script hash includes
the code hash, the hash type (which we just ignored before this) and the args.

The following members are now script hashes:
- payment_script_hash of a participant (specifies which script to use to pay out a participant
- the args of the pfls (this is now the script hash of the type script of the channel to which
	the funds are locked
- unlock_script_hash of a participant

The following member are now code hashes (together with their hash type):
- pfls_code_hash: specifies which pfls to use
- pcls_code_hash: specifies which pcts to use
Here, we do not want to use script hashes, as this would introduce cyclic dependencies of hash
inputs.

The funding and payout checks, as well as the integrity checks on the channel constants were
adjusted accordingly